Add anticaptcha support and change project structure
