+++ tests/test_freepacktbook.py
@@ -1,11 +1,13 @@
 import os
 from sys import version_info
 
-import pytest
-
-from freepacktbook import FreePacktBook, InvalidCredentialsError
 from mock import mock_open, patch
+import pytest
+import requests_mock
 from vcr import VCR
+
+from freepacktbook import FreePacktBook
+from freepacktbook.freepacktbook import InvalidCredentialsError
 
 if version_info.major == 2:
     import __builtin__ as builtins
@@ -16,54 +18,57 @@ else:
 vcr = VCR(path_transformer=VCR.ensure_suffix('.yaml'),
           cassette_library_dir=os.path.join('tests', 'cassettes'))
 
-BOOK_TITLE = 'Multithreading in C# 5.0 Cookbook'
-BOOK_SLUG = 'multithreading_in_c_5_0_cookbook'
-BOOK_ID = '12544'
-
 
 @pytest.fixture
 def packtpub_client():
-    return FreePacktBook('test@example.com', 'mypassword')
+    client = FreePacktBook('test@example.com', 'mypassword')
+    client.session.cookies['SESS_live'] = 'test'
+    return client
 
 
 @vcr.use_cassette
-def test_claim_free_ebook(packtpub_client):
-    book = packtpub_client.claim_free_ebook()
-    assert book['title'] == BOOK_TITLE
-    assert book['id'] == BOOK_ID
+def test_get_book_details(packtpub_client):
+    book = packtpub_client.get_book_details()
+    assert book['title'] == 'Git - Version Control for Everyone'
+    assert book['id'] == '10558'
 
 
 @vcr.use_cassette
 def test_my_books(packtpub_client):
     books = packtpub_client.my_books()
-    assert books[0]['id'] == BOOK_ID
-    assert books[0]['title'] == BOOK_TITLE
+    assert books[0]['id'] == '12544'
+    assert books[0]['title'] == 'Multithreading in C# 5.0 Cookbook'
 
 
-@vcr.use_cassette
 def test_download_book(packtpub_client, monkeypatch):
-    monkeypatch.setattr('freepacktbook.path.exists', lambda x: False)
-    monkeypatch.setattr('freepacktbook.makedirs', lambda x: None)
-    monkeypatch.setattr('freepacktbook.rename', lambda *args: None)
-    book = packtpub_client.claim_free_ebook()
+    monkeypatch.setattr('freepacktbook.freepacktbook.path.exists', lambda x: True)
+    monkeypatch.setattr('freepacktbook.freepacktbook.rename', lambda *args: None)
+    book = {
+        'id': '1234',
+        'title': 'Test Book'}
     with patch.object(builtins, 'open', mock_open()) as result:
-        packtpub_client.download_book(
-            book, destination_dir='/test',formats=['epub'])
-    assert result.call_args[0][0] == '/test/%s/%s.epub_incomplete' % (
-        BOOK_TITLE, BOOK_SLUG,)
+        with requests_mock.mock() as m:
+            url = packtpub_client.download_url % {
+                'book_id': book['id'], 'format': 'epub'}
+            m.get(url, headers={'Content-Length': '100'})
+            packtpub_client.download_book(
+                book, destination_dir='/test',formats=['epub'], override=True)
+    assert result.call_args[0][0] == '/test/Test Book/test_book.epub_incomplete'
 
 
-@vcr.use_cassette
 def test_download_code_files(packtpub_client, monkeypatch):
-    monkeypatch.setattr('freepacktbook.path.exists', lambda x: False)
-    monkeypatch.setattr('freepacktbook.makedirs', lambda x: None)
-    monkeypatch.setattr('freepacktbook.rename', lambda *args: None)
-    book = packtpub_client.claim_free_ebook()
+    monkeypatch.setattr('freepacktbook.freepacktbook.path.exists', lambda x: True)
+    monkeypatch.setattr('freepacktbook.freepacktbook.rename', lambda *args: None)
+    book = {
+        'id': '1234',
+        'title': 'Test Book'}
     with patch.object(builtins, 'open', mock_open()) as result:
-        packtpub_client.download_code_files(
-            book, destination_dir='/test')
-    assert result.call_args[0][0] == '/test/%s/%s_code.zip_incomplete' % (
-        BOOK_TITLE, BOOK_SLUG,)
+        with requests_mock.mock() as m:
+            url = packtpub_client.code_files_url % {'id': int(book['id']) + 1}
+            m.get(url, headers={'Content-Length': '100'})
+            packtpub_client.download_code_files(
+                book, destination_dir='/test', override=True)
+    assert result.call_args[0][0] == '/test/Test Book/test_book_code.zip_incomplete'
 
 
 @vcr.use_cassette
